Fix Joran classloading error if logback 1.3 isn't on classpath
Depending on the JVM, the sanity check that we perform to search
for Joran's model class may not be enough: it is possible that
more classes are eagerly loaded by the VM. To avoid this, we move
the code which is using Joran 1.3 classes into its own class.
